Fixes #8839 change value on response location header
The string does NOT need to be decoded to set it on the location header,
the header is going to be used as a value for a redirect and passed
again through the filters.  If the value comes from the uri override
attribute, used it, if not use the same uri as the current request.