Fixes #8952 Validate binary field as File type
The system tries to save the binary field using the path to a non
existing file on a temporary directory.  The file is not there since
it was never edited, so the map contains the reference to the path 
instead of the File object.  The process previously ignored the error.