#9073 disable display of anonymous user on user admin screen and delete
Removes from the API the possibility to delete the anonymous user
and adds a test for it.  Also updates the methods used to get the list
into the user admin screen, keeping anonymous from the list.  The old
method still works as before referencing the new one with the default
value of includeAnonymous = true.