#9619 Avoid overriding of content's map values.
* ContentletUtil.java: setSpecialFields was removed and replace by a new method since it was messing with some values in the contentlet
that resulted in images and files lost. New method 'getContentPrintableMap' was created to obtain a map
with content's original map as well as tags, binaries and categories. This is now call in places when we need to print the content including
binaries, categories and tags (e.g. Content endpoint, ES Portlet)
* ContentResource.java: modifications to make use of the ContentletUtil.getContentPrintableMap
* render.jsp: modifications to make use of the ContentletUtil.getContentPrintableMap