Fix broken links in `Update Available` popup
Links to the change log for the versions are broken (pointing to both a doc that doesn't exist and an anchor that doesn't exist).  The existing link redirects to the change log (via a Vanity URL), but since the anchor is broken it doesn't take you to the appropriate location in the change log.