dotCMS/core#20342 [Content Types] : Deleting field is throwing an error
As per discussion with Freddy R, we noticed that the API is trying to create relationships in any Relationship field even after simply deleting or moving other fields. The code in charge of moving and deleting fields is now setting the skip Relationship creation parameter to "true" so that this doesn't happen. Also, logging the error when the field layout is invalid, and adding Javadoc. (#20344)