dotCMS/core#19837 Upgrade NX, Angular y PrimeNG (#1842)
* Clean

* Run nx migrate to 12

* Update Storybook to 6.2.X
Because the NX migration was need it in order to use Webpack 5

* Runned: nx migrate --run-migrations

* Fix circular dependecies in dotcms-models

* Fix the build for dotcms-webcomponents

Had to edit the angular.json to add the outputPath in the dotcms-webcomponents due the following error:

> nx run dotcms-webcomponents:build --verbose --no-cache
The "path" argument must be of type string. Received undefined
TypeError [ERR_INVALID_ARG_TYPE]: The "path" argument must be of type string. Received undefined
    at validateString (internal/validators.js:124:11)
    at join (path.js:1039:7)
    at /Users/fmontes/Developer/core-web/node_modules/@nxext/stencil/src/executors/stencil-runtime/stencil-config.js:26:77
    at Generator.next (<anonymous>)
    at /Users/fmontes/Developer/core-web/node_modules/tslib/tslib.js:117:75
    at new Promise (<anonymous>)
    at __awaiter (/Users/fmontes/Developer/core-web/node_modules/tslib/tslib.js:113:16)
    at initializeStencilConfig (/Users/fmontes/Developer/core-web/node_modules/@nxext/stencil/src/executors/stencil-runtime/stencil-config.js:15:34)
    at /Users/fmontes/Developer/core-web/node_modules/@nxext/stencil/src/executors/build/executor.js:36:96
    at Generator.next (<anonymous>)

* Fix dotcms-ui build

* Fix storybook for dotcms-block-editor

* Fix dotcms-ui storybook
Knobs are deprecated: https://github.com/storybookjs/storybook/discussions/15060

* Fix SASS warning to use math.div

* Remove migrations.json file because it gets generated by nx

* Migrate to latest and run-migrations

* Remove dot-site-browser

This module was breaking the angular build for "DOMError" not found.

Since we are not using this module at all instead of fixing this error I just remove the module, if we need something we can get it from the git story

* Unify package.json version caret

* Make Storybook run!

First I had this error and I fixed by "cleaning" nx cache:

rm -rf ./node_modules/.cache/nx

(node:41972) UnhandledPromiseRejectionWarning: TypeError: Cannot read property 'data' of undefined
    at ProjectHasher.<anonymous> (/Users/fmontes/Developer/core-web/node_modules/@nrwl/workspace/src/core/hasher/hasher.js:246:55)
    at Generator.next (<anonymous>)
    at /Users/fmontes/Developer/core-web/node_modules/tslib/tslib.js:117:75
    at new Promise (<anonymous>)
    at __awaiter (/Users/fmontes/Developer/core-web/node_modules/tslib/tslib.js:113:16)
    at /Users/fmontes/Developer/core-web/node_modules/@nrwl/workspace/src/core/hasher/hasher.js:241:93
    at new Promise (<anonymous>)
    at ProjectHasher.<anonymous> (/Users/fmontes/Developer/core-web/node_modules/@nrwl/workspace/src/core/hasher/hasher.js:241:50)
    at Generator.next (<anonymous>)
    at /Users/fmontes/Developer/core-web/node_modules/tslib/tslib.js:117:75
(Use `node --trace-warnings ...` to show where the warning was created)
(node:41972) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). To terminate the node process on unhandled promise rejection, use the CLI flag `--unhandled-rejections=strict` (see https://nodejs.org/api/cli.html#cli_unhandled_rejections_mode). (rejection id: 1)
(node:41972) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.

Then run into this issue. The problem here is that we have multiple webpack versions (some package use v4) so we added to the package.json

    "resolutions": {
        "weboack": "^5.65.0"
    }

(node:42076) UnhandledPromiseRejectionWarning: TypeError: The 'compilation' argument must be an instance of Compilation
    at Function.getCompilationHooks (/Users/fmontes/Developer/core-web/node_modules/@angular-devkit/build-angular/node_modules/webpack/lib/javascript/JavascriptModulesPlugin.js:141:10)
    at SourceMapDevToolModuleOptionsPlugin.apply (/Users/fmontes/Developer/core-web/node_modules/@angular-devkit/build-angular/node_modules/webpack/lib/SourceMapDevToolModuleOptionsPlugin.js:50:27)
    at /Users/fmontes/Developer/core-web/node_modules/@angular-devkit/build-angular/node_modules/webpack/lib/SourceMapDevToolPlugin.js:163:53
    at Hook.eval [as call] (eval at create (/Users/fmontes/Developer/core-web/node_modules/@storybook/builder-webpack5/node_modules/tapable/lib/HookCodeFactory.js:19:10), <anonymous>:297:1)
    at Hook.CALL_DELEGATE [as _call] (/Users/fmontes/Developer/core-web/node_modules/@storybook/builder-webpack5/node_modules/tapable/lib/Hook.js:14:14)
    at Compiler.newCompilation (/Users/fmontes/Developer/core-web/node_modules/@storybook/builder-webpack5/node_modules/webpack/lib/Compiler.js:1055:26)
    at /Users/fmontes/Developer/core-web/node_modules/@storybook/builder-webpack5/node_modules/webpack/lib/Compiler.js:1099:29
    at _next0 (eval at create (/Users/fmontes/Developer/core-web/node_modules/@storybook/builder-webpack5/node_modules/tapable/lib/HookCodeFactory.js:33:10), <anonymous>:23:1)
    at eval (eval at create (/Users/fmontes/Developer/core-web/node_modules/@storybook/builder-webpack5/node_modules/tapable/lib/HookCodeFactory.js:33:10), <anonymous>:35:1)
    at processTicksAndRejections (internal/process/task_queues.js:93:5)
(Use `node --trace-warnings ...` to show where the warning was created)
(node:42076) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). To terminate the node process on unhandled promise rejection, use the CLI flag `--unhandled-rejections=strict` (see https://nodejs.org/api/cli.html#cli_unhandled_rejections_mode). (rejection id: 1)
(node:42076) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.

* Fix the stories path to show all in dotcms-ui storybook

* Run dotcms-webcomponents storybook

At this points run but since knobs were deprecated we need to replace all the stories with controls

https://storybook.js.org/docs/react/essentials/controls

Also we might want to take this apporach:

https://mokkapps.de/blog/run-build-and-deploy-stencil-and-storybook-from-one-repository/

* Make typedoc build dotcms and dotcms-models docs again

* Update workspace to version 2

We were in version 1 of the worspace that uses inline configuration in the angular.json, that's no longer recommended so to do the migration:

1. Update angular.json version to 2
2. Run nx format
3. Run nx g convert-to-nx-project --all

And that did all the changes in this commit to do that migration:

https://nx.dev/configuration/projectjson#workspace-json

* Add vscode settings

* Update .prettierignore

* Update primeng, primeflex and primeicons

* Fix run tests locally

* Update nodejs in tests github action and add engiges to the package.json

* Fix tests round 1

* Fix tests round 2

* Fix tests round 3

* Remove fit