#21786 : Stop sending the "rendered" property in the JSON response for EMA Sites.
As discussed with Will and the Lunik Team, the `EMAWebInterceptor` class will add the "EMA_REQUEST" attribute when when a request goes through it. Then, the `/api/v1/page/render` method will check for that and if it is set, pass the args to the `/api/v1/page/json` method instead, and return its result.