#22116 : DB Exporter must stream output
When we moved the exporter plugin into the core, we're waiting for a `pg_dump` file to be written before we start streaming back to the response. This basically breaks the functionality in production environments and means that when run against an instance behind a gateway or load balancer, the load balancer will close the connection before the file gets generated. We need to stream the output to the response.