#22809 : Adding Upgrade Task to fix the 'Show On Menu' property.
During some customer upgrades, there was a situation in which the value of the `show_on_menu` column in the `contentlet` table didn't match the value present in the `text` column. This new task checks both the legacy `text` column and the new `contentlet_as_json` column to see if the 'Show On Menu' property is enabled. If it is, the `show_on_menu` value is updated to match it.