#18742 : Shallow Push must work as expected. (#23178)
* #18742 : Shallow Push must work as expected.

There are situations in which an HTML Page is shallow-pushed, i.e., none of its child contents are pushed along. If a User opens up that page in the receiving instance, this can result in the MultiTree cache storing an entry with an empty list of child contents. So, the next time any of such contents are pushed, they will not show up in the HTML Page, requiring the MultiTree cache to be flushed. This PR checks this situation, and flushes the respective cache entry when appropriate.

* Fixing more broken Integration Tests.