#22771 : UT Task210901UpdateDateTimezones not running successfully (#23096)
* #22771 : UT Task210901UpdateDateTimezones not running successfully

The `HikariDataSource` is creating `Connection` objects with the default `TimeZone` value -- for example, `America/Costa Rica` -- instead of using the time zone set in the database: UTC. When the `Task210901UpdateDateTimezones` runs and updates the existing date columns to `TIMESTAMP WITH TIMEZONE`, the date values are being transformed into your local Time Zone. This is wrong because this process is NOT supposed to change dates, only the column type, which is what happens when the queries are run directly in Postgres.

The solution is this case is to call `TimeZone.setDefault(TimeZone.getTimeZone("UTC"));` before creating the Data Source instance so that the `Connection` objects use the `UTC` Time Zone instead of the one set in your environment. This way, dates will remain as they are when the column types get updated, and the Upgrade Task will take care of transforming them into the appropriate Time Zone based on the Locale selected by the customer in the dotCMS back-end.

* #22771 : Adding a new test scenario. We now need to check that the `DbConnectionFactory.getConnection()` and the `DbConnectionFactory.getDataSource().getConnection()` methods return `Connection` objects that are always referencing the UTC Time Zone.

* #22771 : As per Erick's request, updating all date column values in the `identifier` table.

* #22771 : Adding one more verification to the IT.

* #22771 : Applying generic solution for Daylight Savings Time. Fixing Integration Test. Adding more details to the error message in the identifier parent path check trigger for MSSQL and Postgres.

* #22771 : Bring the original default Time Zone back after the Data Source instance has been created.