dotCMS/core#22832 Update the image node block to support dot-images and html images
* #23088 : System Container shows broken images when adding.

Updated the `default_container.vtl` file to use the `$dotContentMap` variable instead of calling other ViewTools to get the data we need. Also, fixed a problem in which the  working version of a Contentlet was not being neither added nor shown up in Edit or Preview Mode. This was changed to always get the working version of the Contentlet when in Edit or Preview Mode, and the live version in Live Mode.

* Fixing Integration Test: `ContentUtilsTest.test_find_multilingual_contentlet_getContentletSuccessfully()`