#23223 : Adding page reference counter to Contentlets.
IN EDIT MODE ONLY, we need to let the User choose whether they want to update the "global" Contentlet reference, or if they want to create an exact copy of the Contentlet in that specific Container and page, and edit it. In order to do this, we're adding a new parameter to the Contentlet data map called `onNumberOfPages` which allows the UI to provide Users with this option.

The page reference calculation will be carried out only once, until the very same Contentlet is either added or removed from any HTML Page. A new Cache Group was created in order to keep track of this.