#20529 : Refactor UserResource filter endpoint
Allowing the generation of paginated results for specific `UserResource` endpoints using the expected `PaginationUtil` class. Moreover, the existing functionality that allowed users to filter by Asset Inode and Permission Type was not working at all, so it had to be fixed as well. Adding and fixing Integration Tests. Adding Javadoc.