dotCMS/core#20529 Refactor UserResource filter endpoint
* #20529 : Refactor UserResource filter endpoint

Allowing the generation of paginated results for specific `UserResource` endpoints using the expected `PaginationUtil` class. Moreover, the existing functionality that allowed users to filter by Asset Inode and Permission Type was not working at all, so it had to be fixed as well. Adding and fixing Integration Tests. Adding Javadoc.

* Fixing Unit Test.

* Temporarily ignoring test method.

* Fixing Unit Tests. Removing unnecessary code in the `UserPaginator` class.

* Implementing SonarQube's feedback.

* Implementing SonarQube's feedback.

* Adding missing change reported by failure in Integration Test.

* Fixing `UserResource.postman_collection.json` Postman test.