#24740 Rename file and method names for clarity
Change the method name from `buildFileSystemTree` to `pullTreeNode` and modified the corresponding task's name and description to `PullTreeNodeTask` from `LocalFoldersTreeBuilderTask`. The change is made to more accurately reflect the method/task's functionality which is to pull the contents of a tree node from a remote server rather than just building a local file system tree.