#25661 Add dotCMSObjectType field to language, content type and site models
Added a "dotCMSObjectType" field to the language, content type, and site models in the project to get better clarity. Additionally, new tests have been added to LanguageCommandTest, ContentTypeCommandTest, and SiteCommandTest. These tests ensure that the "dotCMSObjectType" field is present in the JSON and YAML files generated when the pull command is executed for languages, content types, and sites respectively.

This addition helps in identifying the object type when reviewing the JSON or YAML files, improving maintainability and understandability of the codes and files.