Issue 25661 add an attribute to each and every piece of json that conforms the cli bundle (#25731)
* #25661 Add dotCMSObjectType field to language, content type and site models

Added a "dotCMSObjectType" field to the language, content type, and site models in the project to get better clarity. Additionally, new tests have been added to LanguageCommandTest, ContentTypeCommandTest, and SiteCommandTest. These tests ensure that the "dotCMSObjectType" field is present in the JSON and YAML files generated when the pull command is executed for languages, content types, and sites respectively.

This addition helps in identifying the object type when reviewing the JSON or YAML files, improving maintainability and understandability of the codes and files.

* #25661 Refactor ContentType class: "TYPE" made final

* #25661 Added a JsonProperty to ContentType class

* #25661 Added JsonViews and JsonProperty to handle JSON serialization.

This commit introduces JsonViews in AbstractLanguage, ContentType, and AbstractSiteView classes. This helps manage different JSON views and control serialization for different use cases. A JsonProperty "dotCMSObjectType" has been added to the ContentType class. Test cases for Content Type, Site and Language have been updated to include a command to push the changes back to the server, ensuring the structure is still correct. New functionalities for creating and updating language and content type have been provided with the ExternalView of Json, adding more security to the API calls. The CommonViews class has been added to handle different views for JSON serialization. This change enhances the control over data visibility in different scenarios.

* #25661 Cleanup

---------

Co-authored-by: Nollymar Longa <nollymar.longa@dotcms.com>