#24737 Update code to support file assets with spaces
The code has been updated to support file assets with spaces in the name and to encode/decode URIs accordingly. The gitignore file has been updated to ignore .war/ files. Changes were made in several locations including the RemoteTraversalServiceTest.java and LocationUtils.java files to improve the handling of files with spaces in their names.