#24737 Update code to support file assets with spaces (#25783)
* #24737 Update code to support file assets with spaces

The code has been updated to support file assets with spaces in the name and to encode/decode URIs accordingly. The gitignore file has been updated to ignore .war/ files. Changes were made in several locations including the RemoteTraversalServiceTest.java and LocationUtils.java files to improve the handling of files with spaces in their names.

* #24737 Update test folder structure and API usage for handling spaces

Refactored folder structure creation in tests to include a folder with a space in its name to verify the proper handling of such cases. Adjusted API calls in tests to use a new encodePath() method which handles spaces in folder and file names, ensuring a proper URL encoding. The change in the AssetsUtils was needed to handle any spaces in the asset names when creating the remote folder path.

Also, updated references to URLDecoder to use the new encodePath() method, simplifying the process and ensuring consistency. This affects methods such as 'resolveAssetAndFolder' in 'AssetPathResolver.java'.

These changes are necessary for ensuring proper test coverage for folders and files with spaces, and for improving code readability and maintainability.

* Reverting changes

* #24737 Improve tests for handling of assets and folders with spaces in names