#25721 Add logging for executed commands
A new class, `LoggingExecutionStrategy`, has been added to the project. This class implements `IExecutionStrategy` with the added functionality to log the command being executed. It is used in `EntryCommand` class to wrap the existing execution strategy, allowing us to log the commands entered by users in the command-line interface. This helps in debugging by providing a clearer picture of the sequence of commands executed.