Issue 25721 cli improve logging handling (#25858)
* #25721 Logging configurations in application.properties files have been updated to improve overall logging efficiency.

* #25721 Logging configurations in application.properties files have been updated to improve overall logging efficiency.

* #25721 Logging configurations in application.properties files have been updated to improve overall logging efficiency.

* #25721 Logging configurations in application.properties files have been updated to improve overall logging efficiency.

* #25721 Add logging for executed commands

A new class, `LoggingExecutionStrategy`, has been added to the project. This class implements `IExecutionStrategy` with the added functionality to log the command being executed. It is used in `EntryCommand` class to wrap the existing execution strategy, allowing us to log the commands entered by users in the command-line interface. This helps in debugging by providing a clearer picture of the sequence of commands executed.

* #25721 Updating documentation