ci: Add merge_group trigger to allow testing of github merge queue
This will do nothing on its own, but when merge queue is turned on it will run the main workflow with the set of combined commits and is needed to pass before the batch of pull requests is allowed onto master.  We may need to alter some of the reporting to handle this trigger type.