#26383 Add functionality for pulling Content-type descriptors
This commit adds a new feature allowing the retrieval of Content-type descriptors within the dotCMS instance. This functionality fetches and saves the descriptor information for Content-types, retrieving all descriptors by default unless a specific Content-type's name or ID is provided as an argument. Files are saved in JSON by default, but YAML output can be specified, and will overwrite existing descriptor files if the command is performed multiple times. This feature aims to simplify the process of analyzing and working with Content-type data within a dotCMS system.