Issue 26383 cli implement a global pull for content types (#26677)
* #26383 Add functionality for pulling Content-type descriptors

This commit adds a new feature allowing the retrieval of Content-type descriptors within the dotCMS instance. This functionality fetches and saves the descriptor information for Content-types, retrieving all descriptors by default unless a specific Content-type's name or ID is provided as an argument. Files are saved in JSON by default, but YAML output can be specified, and will overwrite existing descriptor files if the command is performed multiple times. This feature aims to simplify the process of analyzing and working with Content-type data within a dotCMS system.

* #26383 Update Content-type pull command and tests

Modified the content-type pull command in the documentation and in test cases to retrieve all the content types by default instead of requiring a specific id or name.

* #26383 Improving test

* #26383 Changed filename to correct case for test

Changed filename from "fileAsset.json" to "FileAsset.json" in ContentTypeCommandIT unit test. The file name is case-sensitive and was causing the test to fail when trying to find and read the file.