Update dotmarketing-config.properties (#26967)
The `json.web.token.allowhttp` property had its default changed from `true` to `false` about a year ago, but the comment above it was not updated to match; it still seemed to imply a default of `true`, and encouraged users to change it in a production environment. This could lead to confusion, or worse!

Co-authored-by: erickgonzalez <erick.gonzalez@dotcms.com>