fix(CLI): 404 Errors When Pulling Content from the CLI (#28032)
* #27985 Removing unnecessary checkins when pushing content from the CLI and also removing unneeded ES dependency in the `getContentUnderParentFromDB` method.

The unnecessary code in the 'WebAssetHelper' method in the 'WebAssetHelper' class has been removed, simplifying the publication and unpublishing of content. In addition, the 'getContentUnderParentFromDB' and 'selectQuery' methods in 'BrowserAPIImpl' class have been refactored to remove the use of Lucene query and focus more on SQL query, hence improving code readability and possibly performance as well.

* #27985 Refactor WebAssetHelper checkinOrPublish method. The code adjustments ensure that an item is checked in and processed only if necessary.

* #27985 Refactor SQL query generation in BrowserAPIImpl to reduce "its Cognitive Complexity" as requested by sonarlint.