ci(report) fix broken report logic attempt 2 (#28288)
Logic as a bit strange it turns out that when there is no input passed it evaluates to null.  we need this case to evaluate a false and only when inputs.slack-only-on-failure is explicitly set to false should the result be true.

!inputs.slack-only-on-failure : failed as not set "null" is flipped to true.
inputs.slack-only-on-failure == "false" breaks as it is a boolean and is not automatically converted to a string.

inputs.slack-only-on-failure == false : matches true when input explicitly set to false, false when the default input value "true" is set, and false when input does not exist as null == false,  it should not error checking null here.