ci(report) only send on failure status attempt #3
Still not working properly to send only with a false status.  Going crazy,  think this must be it but can check my logic.  

An actual check of the workflow trigger is required to distinguish between an overridden default value to be set as false, and the value just not being set which is resolved as false also.  If we are not being called as a reusable workflow with workflow_call trigger then we will need the failure status to run.