#27979 New lang vars endpoint and findAllVariables methods (#28429)
### Proposed Changes
* The main goal of this PR is to introduce the new methods that will be
used to replace the old Language Keys stored as property files
* These methods have been added to LanguageVariableAPI
* The new methods are designed to simply return the LanguageVariables no
user or permission is contemplated here.
* Even though there were a few other methods here designed to return
KeyValue objects these have a problem with being index dependant
* These LanguageVariables are cached using the same LanguageCache but a
separate group. All the Language Variables are cached per language. so
if we update or unpublish a LanguageVariable let's say (and cache
invalidation is needed) This will not affect other LanguageVaribales of
a different language living cache.
* A new Endpoint for retrieving LanguageVariables has been added.
* The respective postman test was added too.  
* Changes were made to ESContentletAPIImpl so the respective cache
invalidation can be accomplished.
* Several issues were outlined on ESContentletAPIImpl by sonar so they
were corrected.