feat(backend): PreJ21 Update to Junit 5 with Vintage Engine Updat (#27995) (#28587)
### Proposed Changes
* Upgrade Mockito
* Remove PowerMock (replace Static mocks with Mockito equivalent)
* Upgrade ByteBuddy
* Upgrade to Unit5 framework with vintage engine for 4 support


Mockito now errors if when clauses are never used, helps to identify
extra clauses that may be expected to be hit but never are and allow for
cleanup if they are not required. There is a Mockito.lenient option that
allows it to be used still.


### Checklist
- [ ] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)

### Additional Info
Related to #27995 (PreJ21 Update to Junit 5 with Vintage Engine Updat).


### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **

---------

Co-authored-by: Will Ezell <will@dotcms.com>