feat(sdk-client): Lucene Query Builder Architecture (#28539)
## Proposed changes
- Add composite/builder class `QueryBuilder` to handle the simple cases
of building a Lucene Query
- Add tests to be sure that the class works as expected

## Lucene Syntax


> [!NOTE]  
I used this
[docs](https://lucene.apache.org/core/2_9_4/queryparsersyntax.html) as
reference for this Query Builder
> 

### Field
Lucene supports fielded data. When performing a search you can either
specify a field, or use the default field. The field names and default
field is implementation specific.

You can search any field by typing the field name followed by a colon
":" and then the term you are looking for.

Example: `+myField:`

### Term
A query is broken up into terms and operators. There are two types of
terms: Single Terms and Phrases.

A Single Term is a single word such as "test" or "hello".

A Phrase is a group of words surrounded by double quotes such as "hello
dolly".

### Operand

Boolean operators allow terms to be combined through logic operators.
Lucene supports AND, "+", OR, NOT and "-" as Boolean operators(Note:
Boolean operators must be ALL CAPS).

By the time, we only support `OR`, `AND` & `NOT`



## Architecture Diagram

I used a combination of composite and builder pattern to create a cycle
that allows the developer to always write a lucene query with a correct
syntax and avoid early building of queries.

<img width="1475" alt="Screenshot 2024-05-14 at 10 45 05 AM"
src="https://github.com/dotCMS/core/assets/63567962/de33cb5d-7e0a-4e7d-a9f6-0035fe158add">

Look how we can only end the cycle by getting to a `term` and calling
the `build` method, this is to prevent the building of invalid Queries.

We can say that a Lucene Query should always end in a Term to be valid
and complete.

Update: Term naming was changed to Equals in the code.