Issue 28251 replacing api calls  (#28544)
### Proposed Changes
* A Feature Flag in the form of a method called
isLocalizationEnhancementsEnabled has been added to activate/deactivate
the functionality.
* The Feature flag is for now on. If we want to release we need to set
it to off.
* Added the capability of accessing the language variable in the cache
using only the key and the lang ID
* Fixed a bug in the queries introduced when changed to show both live
and working content. The bug was making all archived pieces of content
appear in the results.
* Fixes on Integration Tests
* Fixes on postman test. That was caused by a dupe key introduced during
the postman run. The unique key constraint can be lifted, so we must be
prepared for such a situation.
* New API Calls added next to the old one. The new ones are executed
instead of the old APIs only if the feature flag is on.