#28588 Create API Method to get Experiment by Host (#28630)
second PR for:

https://github.com/dotCMS/core/issues/28588

first one: https://github.com/dotCMS/core/pull/28628

Refactor MultiTreeCache to make it cluster-aware.

Previously, we used the Page ID as the key in the MultiTreeCache,
storing a map where each key represented a different variant name and
each value was a MultiTree. When we needed to remove a variant, we
followed these steps:

Removed the item from the map representing the variant.
Put the updated map back into the cache.
However, the put method was not cluster-aware. You can see the relevant
code here:


https://github.com/dotCMS/core/blob/eb200d27482a557634e0f732ed5f055c0df28450/dotCMS/src/main/java/com/dotmarketing/cache/MultiTreeCache.java#L91-L103

### Proposed Changes
* Create the listActive Experiment API Method to get all the Active
Experiment by Host


https://github.com/dotCMS/core/pull/28630/files#diff-8bf1865184fd59c5c64882df72bd0c71a403a544777830acf250bbf54ea84d90R271

* Use Page's id + variant Name as Cache key 


https://github.com/dotCMS/core/pull/28630/files#diff-ad644a666145836b324394468b653fe572daf6de9d1e434752ec44eca7fdf9a4R53

so we don't need the Map anymore

* When we need to remove all the variants for a Page, we use the new
listActive API Method


https://github.com/dotCMS/core/pull/28630/files#diff-ad644a666145836b324394468b653fe572daf6de9d1e434752ec44eca7fdf9a4R99

### Checklist
- [x] Tests