Issue 28596 use add-opens and set --illegal-access=deny in preparation for Java 21 (#28619)
### Proposed Changes

- Fix tika api versioning and tie it to build/release
- Add required add-opens parameters to work around illegal access issues
and set set —illegal-access=deny to replicate J21 behavior
- Make ContentTypeUtilsTest succeed regardless of the parameter order
returned which is JVM dependent.
- Add junit-jupiter-suite and junit-platform-suite dependencies to
prepare for Junit4 → Junit 5 migration of Current Suite classes and
DataProvider test cases
- Normalize some duplicate definitions of Java version to reduce changes
when a change is made.
- Upgrade dnsjava to 3.5.3 that supports java 21 without extra
add-opens.

Tests: Should focus on any issues with bundles due to the serialization
and deserialization and errors in logs indicating there is illegal
access to internal classes. Serialization during caching would also be a
concern area and reporting of cache sizes.



### Checklist
- [ ] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)

### Additional Info
** any additional useful context or info **

### Screenshots
Original             |  Updated
:-------------------------:|:-------------------------:
** original screenshot **  |  ** updated screenshot **