feat(CLI): Watch Mode    Refs:#28593  (#28710)
### Proposed Changes
* We're introducing Watch options in the Push Mixin so they become
available to every push subcommand
* We're adding a bit of logic in the `DotExecutionStrategy` to be able
to handle starting up in watch mode and continue to repeat execution of
the command of choice
* I'm making `DotPush` a descendant of `DotCommand` to help me select
all Push commands and still gain access to the mixins.
* Because of the above change I needed to tell the global Push command
from the others, so I introduced a new getter with such info.
* Symmetry-wise wise I made `DotPull` also a descendant of `DotCommand`.
But just for symmetry. All Pull sub commands got modified when I removed
the no longer-needed DotCommand interface. But there are no other
changes in logic.
* A DirectoryWatchService was introduced. this class uses a poll
interval to observe and filter events.
* The events captured by the service are filtered so, let's say a folder
move that in reality creates between 4 and 5 events (FOLDER_MODIFY,
FOLDER_DELETE, FOLDER_MODIFY, FOLDER_CREATE), These events are filtered
so that we only get the most recent one. This ensures we do not flood
the server side with many unneeded changes.
* Changes in documentation were made
* Tests were introduced for the `DirectoryWatchService`, The Global Push
running with the watch mode flag on, and finally another one to cover
the changes introduced in the `DotExecutionStrategy`