build: Use consistent java versions using sdkman (#28868)
### Proposed Changes
* Change java base image to use sdkman to get java version. Use
SDKMAN_JAVA_VERSION build arg to specify SDKMAN version and
implementation.
* Create manual workflow specifically for creating java base image based
upon SDKMAN version string being tagged and pushed with that version.
* Modify maven to pick up version specified in ~/.sdkmanrc and use that
by default to specify which upstream java base image to used.
* Cleanup places where java version is specified in maven pom files to
default to a single location in parent/pom.xml. Set default value in
environments/environment.properties so it can be modified locally be
developers easily and changed in one place. setting
java.compat.version=11
* Use independent version for compatibility of build allowing Java 21 to
be upgraded to but still allowing plugins to be build with Java 11 with
easy switch later.
* Cleanup DockerFiles with some AI help.


Note: 
* This change does not actually update to java 21 yet. We should do
internal test to upgrade to latest point release of java 11 first. e.g.
11.0.22-ms -> 11.0.23-ms
* This change does not modify the version used in the ci build
workflows. We will follow up a change to enable consistent use of the
.sdkmanrc in the workflows after this PR.