build: curl accidentally removed from docker image (#28918) (#28921)
### Proposed Changes
* Update packages like curl missing from final image that were previousl
in the base.
* core-web/.nvmrc file is re-generated from the build every time but the
file in master currently has an extra new line which is removed on each
build. Including this just to fix it so it does not get added to others
changelists going forward.

I will restructure this after this fix to improve the base image build
and reduce the final image size.

Test that curl is available in the resulting image and that a custom
starter can be specified and used