Issue 28265 cli support sending workflow names when saving content types (#28885)
Now, when updating the content type workflows, we don't depend solely on
the workflow ID, but we can use now the workflow variable name as a
secondary key if needed.

To accomplish this, we needed to introduce in the workflow scheme the
concept of the variable name (Like the variable we have in content types
and fields, same concept).