feat(CICD): Support for issue edited event to notify UI/UX (#29068)
Introducing handling of `Issue edition` event specifically for when the
`Technology` field is set to `Font-end`.
Also, not including an entry point just for this. Instead I'm providing
a general purpose workflow for when an issue is changed and from there
evaluate the actions it needs to execute. Also identified some logic
that was extracted in order to be reused instead of copy-paste for
resolving slack user ids from github accounts and the creation of teams
where github accounts are grouped under a team.